+++ frontend/src/app/modules/shared/components/select/select.component.spec.ts
@@ -19,6 +19,7 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
+import { FormControl } from '@angular/forms';
 import { SelectComponent, SelectOption } from '@shared/components/select/select.component';
 import { SharedModule } from '@shared/shared.module';
 import { screen, waitFor } from '@testing-library/angular';
@@ -26,92 +27,90 @@ import { renderComponent } from '@tests/test-render.utils';
 
 describe('SelectComponent', () => {
   it('should render the select component', async () => {
-    const lable = 'Test';
-    await renderComponent(`<app-select [lable]='lable'></app-select>`, {
+    const label = 'Test';
+    await renderComponent(`<app-select [label]='label'></app-select>`, {
       imports: [SharedModule],
       declarations: [SelectComponent],
-      componentProperties: { lable },
+      componentProperties: { label },
     });
 
-    expect(screen.getByText(lable)).toBeInTheDocument();
+    expect(screen.getByText(label)).toBeInTheDocument();
   });
 
   it('should render the select component with options', async () => {
-    const lable = 'Test';
-    const options: SelectOption[] = [{ lable: 'Test_01' }, { lable: 'Test_02' }];
-    const fixture = await renderComponent(`<app-select [lable]='lable' [options]='options'></app-select>`, {
+    const label = 'Test';
+    const options: SelectOption[] = [{ label: 'Test_01' }, { label: 'Test_02' }];
+    const fixture = await renderComponent(`<app-select [label]='label' [options]='options'></app-select>`, {
       imports: [SharedModule],
       declarations: [SelectComponent],
-      componentProperties: { lable, options },
+      componentProperties: { label, options },
     });
 
-    const selectComponent = screen.getByText(lable);
+    const selectComponent = screen.getByText(label);
     expect(selectComponent).toBeInTheDocument();
     selectComponent.click();
     fixture.detectChanges();
 
-    await waitFor(() => expect(screen.getByText(options[0].lable)).toBeInTheDocument());
-    await waitFor(() => expect(screen.getByText(options[1].lable)).toBeInTheDocument());
+    await waitFor(() => expect(screen.getByText(options[0].label)).toBeInTheDocument());
+    await waitFor(() => expect(screen.getByText(options[1].label)).toBeInTheDocument());
   });
 
   it('should select correct option if selected value is set', async () => {
-    const lable = 'Test';
-    const options: SelectOption[] = [{ lable: 'Test_01' }, { lable: 'Test_02' }];
-    await renderComponent(
-      `<app-select [lable]='lable' [options]='options' [selectedValue]='options[0].lable'></app-select>`,
-      {
-        imports: [SharedModule],
-        declarations: [SelectComponent],
-        componentProperties: { lable, options },
-      },
-    );
+    const label = 'Test';
+    const options: SelectOption[] = [{ label: 'Test_01' }, { label: 'Test_02' }];
+    const formControl = new FormControl(options[1].label);
 
-    const selectComponent = screen.getByText(lable);
+    await renderComponent(`<app-select [label]='label' [options]='options' [formControl]='formControl'></app-select>`, {
+      imports: [SharedModule],
+      declarations: [SelectComponent],
+      componentProperties: { label, options, formControl },
+    });
+
+    const selectComponent = screen.getByText(label);
     expect(selectComponent).toBeInTheDocument();
 
-    await waitFor(() => expect(screen.getByText(options[0].lable)).toBeInTheDocument());
+    await waitFor(() => expect(screen.getByText(options[1].label)).toBeInTheDocument());
   });
 
-  it('should select correct option with value lable mapping if selected value is set', async () => {
-    const lable = 'Test';
+  it('should select correct option with value label mapping if selected value is set', async () => {
+    const label = 'Test';
     const options: SelectOption[] = [
-      { lable: 'Test_01', value: '111' },
-      { lable: 'Test_02', value: '222' },
+      { label: 'Test_01', value: '111' },
+      { label: 'Test_02', value: '222' },
     ];
-    await renderComponent(
-      `<app-select [lable]='lable' [options]='options' [selectedValue]='options[1].value'></app-select>`,
-      {
-        imports: [SharedModule],
-        declarations: [SelectComponent],
-        componentProperties: { lable, options },
-      },
-    );
+    const formControl = new FormControl(options[1].value);
 
-    const selectComponent = screen.getByText(lable);
+    await renderComponent(`<app-select [label]='label' [options]='options' [formControl]='formControl'></app-select>`, {
+      imports: [SharedModule],
+      declarations: [SelectComponent],
+      componentProperties: { label, options, formControl },
+    });
+
+    const selectComponent = screen.getByText(label);
     expect(selectComponent).toBeInTheDocument();
 
-    await waitFor(() => expect(screen.getByText(options[1].lable)).toBeInTheDocument());
+    await waitFor(() => expect(screen.getByText(options[1].label)).toBeInTheDocument());
   });
 
   it('should render the select component with custom option renderer', async () => {
-    const lable = 'Test';
-    const options: SelectOption[] = [{ lable: 'Test_01' }, { lable: 'Test_02' }];
+    const label = 'Test';
+    const options: SelectOption[] = [{ label: 'Test_01' }, { label: 'Test_02' }];
     const fixture = await renderComponent(
       `<ng-template #test let-value='value'>_TEST_{{value}}</ng-template>
-                <app-select [lable]='lable' [options]='options' [optionsRenderer]='test'></app-select>`,
+                <app-select [label]='label' [options]='options' [optionsRenderer]='test'></app-select>`,
       {
         imports: [SharedModule],
         declarations: [SelectComponent],
-        componentProperties: { lable, options },
+        componentProperties: { label, options },
       },
     );
 
-    const selectComponent = screen.getByText(lable);
+    const selectComponent = screen.getByText(label);
     expect(selectComponent).toBeInTheDocument();
     selectComponent.click();
     fixture.detectChanges();
 
-    await waitFor(() => expect(screen.getByText('_TEST_' + options[0].lable)).toBeInTheDocument());
-    await waitFor(() => expect(screen.getByText('_TEST_' + options[1].lable)).toBeInTheDocument());
+    await waitFor(() => expect(screen.getByText('_TEST_' + options[0].label)).toBeInTheDocument());
+    await waitFor(() => expect(screen.getByText('_TEST_' + options[1].label)).toBeInTheDocument());
   });
 });