+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - Added validation for UpdateInvestigationRequest, reason for (decline, accepted, close)
 - Added Title to sections that might be cut of with three dots (...)
+- Added hint to input field with additional information for max and min length validation
 
 ### Changed
 - Restructured helm charts to an parent helm chart which includes frontend and backend
@@ -19,6 +20,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Removed quality type from ui but kept base functionality for future use
 - Improved and separated *.md documentation, especially README.md
 - Improved error page styles and fixed menu urls in terms of 404 pages
+- Improved visibility for field validation and added * to required fields
+- Improved validation error message for max and min length
 
 ## [3.0.0] - 2023-03-21